pyproject.toml
[project]
name = "adk-gdg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.78.7",
    "python-dotenv>=1.1.1",
]
